aracid Long düzeltildi

+++ src/java/dao/HasarKaydiDAO.java
@@ -22,7 +22,7 @@ public class HasarKaydiDAO extends Dao {
                 HasarKaydi tmp;
                 tmp = new HasarKaydi(rs.getInt("hasarid"), rs.getInt("aracid"), rs.getString("boya"), rs.getString("cizik"), rs.getString("degisim"), rs.getString("aciklama"));
 
-                tmp.setArac(this.getAracDAO().find(rs.getInt("aracid")));
+                tmp.setArac(this.getAracDAO().find(rs.getLong("aracid")));
                 clist.add(tmp);//Her yeni hasarKaydii listeme ekliyorum
 
             }
@@ -118,7 +118,7 @@ public class HasarKaydiDAO extends Dao {
                 HasarKaydi tmp;
                 tmp = new HasarKaydi(rs.getInt("hasarid"), rs.getInt("aracid"), rs.getString("boya"), rs.getString("cizik"), rs.getString("degisim"), rs.getString("aciklama"));
 
-                tmp.setArac(this.getAracDAO().find(rs.getInt("aracid")));
+                tmp.setArac(this.getAracDAO().find(rs.getLong("aracid")));
                 clist.add(tmp);//Her yeni hasarKaydii listeme ekliyorum
 
             }
